+++ app/pages/admin/grantbot/components/retirement-rules.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
   
-function retirementRules({ data }) {
+function RetirementRules({ data }) {
   return (
     <section>
       <h3>Have default retirement limits been changed?</h3>
@@ -15,4 +15,12 @@ function retirementRules({ data }) {
   );
 }
 
-export default retirementRules;
+RetirementRules.propTypes = {
+  data: React.PropTypes.arrayOf(React.PropTypes.shape({
+    id: React.PropTypes.string.isRequired,
+    name: React.PropTypes.string.isRequired,
+    retirementHasBeenChanged: React.PropTypes.bool.isRequired,
+  })),
+};
+
+export default RetirementRules;
