+++ app/pages/admin/grantbot/index.js
@@ -1,0 +1,3 @@
+import GrantbotContainer from './containers/grantbot-container';
+
+export default GrantbotContainer;
